+++ packages/core/src/index.ts
@@ -4,3 +4,14 @@ export * from './types/user.type';
 export * from './types/auth';
 export * from './services/cluster';
 export * from './configuration';
+
+export * from './types/asset.evm.v2';
+export * from './types/asset.solana.v2';
+export * from './types/metadata.common.v2';
+export * from './types/metadata.evm.v2';
+export * from './types/metadata.solana.v2';
+export * from './types/wallet.evm.v2';
+export * from './types/wallet.solana.v2';
+export * from './types/nft.v2';
+export * from './types/nft';
+export * from './types/marketplace';