+++ examples/nuxt-ts/CHANGELOG.md
@@ -1,4 +1,108 @@
 # nuxt-example
+
+## 1.2.0
+
+### Patch Changes
+
+- Add SUI page and 4 wallet APIs of SUI
+- Updated dependencies
+  - @mirrorworld/web3.js@2.3.0
+
+## 1.1.2
+
+### Patch Changes
+
+- Add token metadata types"
+- Updated dependencies
+  - @mirrorworld/web3.js@2.2.2
+
+## 1.1.1
+
+### Patch Changes
+
+- Updated dependencies [5ab9055]
+  - @mirrorworld/web3.js@2.2.1
+
+## 1.1.0
+
+### Minor Changes
+
+- 113c546: update @mirrorworld/web3.js package
+
+### Patch Changes
+
+- Updated dependencies [113c546]
+  - @mirrorworld/web3.js@2.2.0
+
+## 1.0.7
+
+### Patch Changes
+
+- Updated dependencies [e2c2c11]
+  - @mirrorworld/web3.js@2.1.0
+
+## 1.0.6
+
+### Patch Changes
+
+- Fixes types for the `QueryEVMNFTsPayloadV2` to support enumeration and range query types
+- Updated dependencies
+  - @mirrorworld/web3.js@2.0.6
+
+## 1.0.5
+
+### Patch Changes
+
+- Fix types exports for marketplace methods"
+- Updated dependencies
+  - @mirrorworld/web3.js@2.0.5
+
+## 1.0.4
+
+### Patch Changes
+
+- af76af2: Fixes types export for Asset service
+- Updated dependencies [af76af2]
+  - @mirrorworld/web3.js@2.0.4
+
+## 1.0.3
+
+### Patch Changes
+
+- Updated dependencies [2781394]
+  - @mirrorworld/web3.js@2.0.3
+
+## 1.0.2
+
+### Patch Changes
+
+- Updated dependencies [11f83a8]
+  - @mirrorworld/web3.js@2.0.2
+
+## 1.0.1
+
+### Patch Changes
+
+- Updated dependencies
+  - @mirrorworld/web3.js@2.0.1
+
+## 1.0.0
+
+### Major Changes
+
+- 5b21017: Migrate SDK to v2
+
+### Patch Changes
+
+- Updated dependencies [5b21017]
+- Updated dependencies [3ce93f7]
+- Updated dependencies [6307447]
+- Updated dependencies [c0a429c]
+- Updated dependencies [50d2d5f]
+- Updated dependencies [a8cb205]
+- Updated dependencies [b5b5e02]
+  - @mirrorworld/web3.js@2.0.0
+
 ## 1.0.0-v2.0
 
 ### Major Changes
@@ -7,11 +111,10 @@
 
 ### Patch Changes
 
-
 - Updated dependencies [3ce93f7]
 - Updated dependencies [6307447]
 - Updated dependencies [c0a429c]
 - Updated dependencies [50d2d5f]
 - Updated dependencies [a8cb205]
 - Updated dependencies [b5b5e02]
-  - @usemirrorworld/web3.js@2.0.0-v2.0
+  - @mirrorworld/web3.js@2.0.0-v2.0