+++ packages/core/CHANGELOG.md
@@ -1,8 +1,111 @@
-# @usemirrorworld/web3.js
+# @mirrorworld/web3.js
+
+## 2.6.0
+
+### Minor Changes
+
+- 181a960: Added 3 APIs onto SUI chain.
+
+## 2.5.0
+
+### Minor Changes
+
+- 5f882b5: Added APIs under 4 Asset functional modules.
+
+## 2.4.0
+
+### Minor Changes
+
+- 225dfd2: Added support for the SUI chain and implemented four interfaces for the wallet module on the SUI chain.
+
+## 2.3.0
+
+### Patch Changes
+
+- Added support for the SUI chain and implemented four interfaces for the wallet module on the SUI chain.
+
+## 2.2.2
+
+### Patch Changes
+
+- Add token metadata types"
+
+## 2.2.1
+
+### Patch Changes
+
+- 5ab9055: Fix return types
+
+## 2.2.0
+
+### Minor Changes
+
+- 113c546: update @mirrorworld/web3.js package
+
+## 2.1.0
+
+### Minor Changes
+
+- e2c2c11: Make `useCredentials` public
+
+## 2.0.6
+
+### Patch Changes
+
+- Fixes types for the `QueryEVMNFTsPayloadV2` to support enumeration and range query types
+
+## 2.0.5
+
+### Patch Changes
+
+- Fix types exports for marketplace methods"
+
+## 2.0.4
+
+### Patch Changes
+
+- af76af2: Fixes types export for Asset service
+
+## 2.0.3
+
+### Patch Changes
+
+- 2781394: Remove `clusterEnv` property from changelog1
+
+## 2.0.2
+
+### Patch Changes
+
+- 11f83a8: Deprecate `env` property from constructir options"
+
+## 2.0.1
+
+### Patch Changes
+
+- Export types"
+
+## 2.0.0
+
+### Major Changes
+
+- 5b21017: Migrate SDK to v2
+- 50d2d5f: Create Multi-chain configurration"
+
+### Minor Changes
+
+- 3ce93f7: Add v2 api client
+- 6307447: Add EVM endpoints helpers and validators
+- b5b5e02: chore: remove loclatoragekey for refresh token on logout"
+
+### Patch Changes
+
+- c0a429c: Fix popup event listeners for close
+- a8cb205: Fix: Logout method - path uri
 
 ## 2.0.0-v2.0
 
 ### Major Changes
+
 - Migrate SDK to v2
 - 50d2d5f: Create Multi-chain configurration"
 
@@ -39,7 +142,7 @@
 
 ### Patch Changes
 
-- 3e0d989: Fix: release new tag on @usemirrorworld/namespace
+- 3e0d989: Fix: release new tag on @mirrorworld/namespace
 
 ## 1.3.1