+++ packages/frameworks/react/CHANGELOG.md
@@ -1,4 +1,34 @@
-# @usemirrorworld/react
+# @mirrorworld/react
+
+## 3.1.0
+
+### Minor Changes
+
+- 113c546: update @mirrorworld/web3.js package
+
+## 3.0.3
+
+### Patch Changes
+
+- Fixes types for the `QueryEVMNFTsPayloadV2` to support enumeration and range query types
+
+## 3.0.2
+
+### Patch Changes
+
+- Fix types exports for marketplace methods"
+
+## 3.0.1
+
+### Patch Changes
+
+- af76af2: Fixes types export for Asset service
+
+## 3.0.0
+
+### Major Changes
+
+- 5b21017: Migrate SDK to v2
 
 ## 3.0.0-v2.0